requirements.txt
# Core ML/DL Libraries
torch==2.1.0
torchvision==0.16.0
timm==0.9.12

# Data Processing
pandas==2.1.3
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
opencv-python==4.8.1.78

# Gradient Boosting
xgboost==2.0.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
grad-cam==1.4.8

# Metrics & Evaluation
scikit-image==0.22.0

# Progress Bars
tqdm==4.66.1

# Jupyter (optional, for notebooks)
jupyter==1.0.0
ipykernel==6.27.1

# Utilities
pyyaml==6.0.1